Refs #32036. Allow sequential fitting in basic and general widgets.
