+++ scripts/Muon/GUI/Common/fitting_widgets/general_fitting/general_fitting_model.py
@@ -449,7 +449,7 @@ class GeneralFittingModel(BasicFittingModel):
         """Returns the runs and group/pairs for the selected data in simultaneous fit by runs mode."""
         runs = self._get_selected_runs()
         groups_and_pairs = [get_group_or_pair_from_name(name) for name in self.fitting_context.dataset_names]
-        workspace_names = [";".join(self.get_fit_workspace_names_from_groups_and_runs([run], groups_and_pairs))
+        workspace_names = ["/".join(self.get_fit_workspace_names_from_groups_and_runs([run], groups_and_pairs))
                            for run in runs]
         return workspace_names, runs, [";".join(groups_and_pairs)] * len(runs)
 
@@ -458,7 +458,7 @@ class GeneralFittingModel(BasicFittingModel):
         runs = [get_run_numbers_as_string_from_workspace_name(name, self.context.data_context.instrument)
                 for name in self.fitting_context.dataset_names]
         groups_and_pairs = self._get_selected_groups_and_pairs()
-        workspace_names = [";".join(self.get_fit_workspace_names_from_groups_and_runs(runs, [group_and_pair]))
+        workspace_names = ["/".join(self.get_fit_workspace_names_from_groups_and_runs(runs, [group_and_pair]))
                            for group_and_pair in groups_and_pairs]
         return workspace_names, [";".join(runs)] * len(groups_and_pairs), groups_and_pairs
 
@@ -479,3 +479,75 @@ class GeneralFittingModel(BasicFittingModel):
     def _update_fit_function_parameters_for_simultaneous_fit(self, _: list, parameter_values: list) -> None:
         """Updates the function parameters for the given dataset names if in simultaneous fit mode."""
         self._set_fit_function_parameter_values(self.fitting_context.simultaneous_fit_function, parameter_values)
+
+    def perform_sequential_fit(self, workspaces: list, parameter_values: list, use_initial_values: bool = False):
+        """Performs a sequential fit of the workspace names provided for the current fitting mode.
+
+        :param workspaces: A list of lists of workspace names e.g. [[Row 1 workspaces], [Row 2 workspaces], etc...]
+        :param parameter_values: A list of lists of parameter values e.g. [[Row 1 params], [Row 2 params], etc...]
+        :param use_initial_values: If false the parameters at the end of each fit are passed on to the next fit.
+        """
+        fitting_func = self._get_sequential_fitting_func_for_normal_fitting_mode()
+        if not self.fitting_context.simultaneous_fitting_mode:
+            workspaces = self._flatten_workspace_names(workspaces)
+        return self._perform_sequential_fit_using_func(fitting_func, workspaces, parameter_values, use_initial_values)
+
+    def _do_sequential_simultaneous_fits(self, row_index: int, workspace_names: list, parameter_values: list,
+                                         functions: list, use_initial_values: bool = False):
+        """Performs a number of simultaneous fits, sequentially."""
+        simultaneous_function = functions[row_index - 1].clone() if not use_initial_values and row_index >= 1 else \
+            self._get_simultaneous_function_with_parameters(parameter_values)
+
+        params = self._get_parameters_for_simultaneous_fit(workspace_names, simultaneous_function)
+
+        return self._do_simultaneous_fit(params, self.fitting_context.global_parameters)
+
+    def _get_simultaneous_function_with_parameters(self, parameter_values: list) -> IFunction:
+        """Returns the current simultaneous function but with the parameter values provided."""
+        simultaneous_function = self.fitting_context.simultaneous_fit_function.clone()
+        self._set_fit_function_parameter_values(simultaneous_function, parameter_values)
+        return simultaneous_function
+
+    def _get_sequential_fitting_func_for_normal_fitting_mode(self):
+        """Returns the fitting func to use when performing a fit in normal fitting mode."""
+        if self.fitting_context.simultaneous_fitting_mode:
+            return self._do_sequential_simultaneous_fits
+        else:
+            return self._do_sequential_fit
+
+    def _update_fit_functions_after_sequential_fit(self, workspaces: list, functions: list) -> None:
+        """Updates the fit functions after a sequential fit has been run on the Sequential fitting tab."""
+        if self.fitting_context.simultaneous_fitting_mode:
+            self._update_simultaneous_fit_function_after_sequential(workspaces, functions)
+        else:
+            super()._update_fit_functions_after_sequential_fit(workspaces, functions)
+
+    def _update_simultaneous_fit_function_after_sequential(self, workspaces: list, functions: list) -> None:
+        """Updates the single fit functions after a sequential fit has been run on the Sequential fitting tab."""
+        for fit_index, workspace_names in enumerate(workspaces):
+            if self._are_same_workspaces_as_the_datasets(workspace_names):
+                self.fitting_context.simultaneous_fit_function = functions[fit_index]
+                break
+
+    def _are_same_workspaces_as_the_datasets(self, workspace_names: list) -> bool:
+        """Checks that the workspace names provided are all in the loaded dataset names."""
+        not_in_datasets = [name for name in workspace_names if name not in self.fitting_context.dataset_names]
+        return False if len(not_in_datasets) > 0 else True
+
+    def _update_fit_statuses_and_chi_squared_after_sequential_fit(self, workspaces, fit_statuses, chi_squared_list):
+        """Updates the fit statuses and chi squared after a sequential fit."""
+        if self.fitting_context.simultaneous_fitting_mode:
+            self._update_fit_statuses_and_chi_squared_for_simultaneous_mode(workspaces, fit_statuses, chi_squared_list)
+        else:
+            super()._update_fit_statuses_and_chi_squared_after_sequential_fit(workspaces, fit_statuses, chi_squared_list)
+
+    def _update_fit_statuses_and_chi_squared_for_simultaneous_mode(self, workspaces: list, fit_statuses: list,
+                                                                   chi_squared_list: list) -> None:
+        """Updates the fit statuses and chi squared after a sequential fit when in simultaneous fit mode."""
+        for fit_index, workspace_names in enumerate(workspaces):
+            if self._are_same_workspaces_as_the_datasets(workspace_names):
+                for workspace_name in workspace_names:
+                    dataset_index = self.fitting_context.dataset_names.index(workspace_name)
+                    self.fitting_context.fit_statuses[dataset_index] = fit_statuses[fit_index]
+                    self.fitting_context.chi_squared[dataset_index] = chi_squared_list[fit_index]
+                break
